// web/platform-admin-web/src/api/generated/model/verifyJoinTermItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { VerifyJoinTermItemScope } from './verifyJoinTermItemScope';

export interface VerifyJoinTermItem {
  /** 약관 문서 식별자 */
  documentId: string;
  /** 약관 버전 식별자 */
  versionId: string;
  /**
     * 조직 식별자
     * @nullable
     */
  organizationId: string | null;
  /** 약관 적용 범위 */
  scope: VerifyJoinTermItemScope;
  /** 필수 동의 여부 */
  required: boolean;
  /** 약관 제목 */
  title: string;
  /** 약관 버전 태그 */
  version: string;
  /** 약관 내용 */
  content: string;
}