// web/platform-admin-web/src/api/generated/model/verifyJoinTermItemScope.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 약관 적용 범위
 */
export type VerifyJoinTermItemScope = typeof VerifyJoinTermItemScope[keyof typeof VerifyJoinTermItemScope];


export const VerifyJoinTermItemScope = {
  platform: 'platform',
  organization: 'organization',
} as const;